add multiqueue pool

+++ main/ColorPalette.cpp
@@ -6,9 +6,9 @@ using mypair = std::pair<float, RGB>;
 
  RGB RGB::interpolate(RGB& low, RGB& high, float lpos, float hpos, float t) {
     float i = (t - lpos) / (hpos - lpos);
-    RGB result = {(high.R - low.R)*i + low.R,
-                  (high.G - low.G)*i + low.G,
-                  (high.B - low.B)*i + low.B};
+    RGB result = {uint8_t((high.R - low.R)*i + low.R),
+                  uint8_t((high.G - low.G)*i + low.G),
+                  uint8_t((high.B - low.B)*i + low.B)};
     return result;
 }
